phase_3/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Todo Backend API - FastAPI + SQLModel + Neon PostgreSQL"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.1",
    "mcp[cli]>=1.26.0",
    "openai-agents>=0.8.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.32",
    "uvicorn>=0.40.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
